==> 1_code/7_discuss_loggedevents.Rmd <==
---
title: "Candidacy: Discuss model output"
author: "Julianna Calabrese"
date: "February 26th, 2023"
output: html_document
---

# Prepwork

```{r}
set.seed(1227)
setwd("/Users/Julianna/OneDrive - The Ohio State University/candidacy/data_analysis")
```

```{r}
#install.packages("Rtools")
#install.packages(c('devtools', 'curl'))
#library('devtools')
#devtools::install_github("simsem/semTools/semTools")
#install.packages("lavaan", repos = "http://www.da.ugent.be", type = "source")
```

```{r}
library(semTools)
library(lavaan)
library(tidyverse)
library(simsem)
library(mice)
library(miceadds)
library(semPlot)
library(lavaanPlot)
library(officer)
library(psych)
library(rrtable)

library(ggmice)
library(micemd)
library(gtsummary)
library(expss)
library(flextable)
library(devtools)
#install_github("jrcalabrese/missy")
library(missy)
library(docxtractr)
library(janitor)# Prepwork
```

# APIM

```{r}
# Small
load("0_data/processed_apim/small.mi1/small.mi1.Rdata")
small.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_apim/small.mi2/small.mi2.Rdata")
small.mi2 <- mi.res
rm(mi.res)

load("0_data/processed_apim/small.mi3/small.mi3.Rdata")
small.mi3 <- mi.res
rm(mi.res)

# Medium
load("0_data/processed_apim/medium.mi1/medium.mi1.Rdata")
medium.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_apim/medium.mi2/medium.mi2.Rdata")
medium.mi2 <- mi.res
rm(mi.res)

load("0_data/processed_apim/medium.mi3/medium.mi3.Rdata")
medium.mi3 <- mi.res
rm(mi.res)

# Large
load("0_data/processed_apim/large.mi1/large.mi1.Rdata")
large.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_apim/large.mi2/large.mi2.Rdata")
large.mi2 <- mi.res
rm(mi.res)

load("0_data/processed_apim/large.mi3/large.mi3.Rdata")
large.mi3 <- mi.res
rm(mi.res)
```

## Logged Events

### Small

```{r}
x <- small.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- small.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- small.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
```

### Medium

```{r}
x <- medium.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- medium.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- medium.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
```

### Large

```{r}
x <- large.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- large.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- large.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
table(x$dep)
```

## Other

```{r}
x <- mice::complete(small.mi1, "long", FALSE) %>%
  select(contains("item")) %>% names
mice_cor(small.mi1, vs = x, title = "hahahaha")
```

```
x %>%
  pivot_longer(cols = everything()) %>%
  ggplot(aes(x = name, y = name, fill = value)) +
  geom_tile()
```

## Meth

```{r}
x <- as.data.frame(small.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(small.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(small.mi3$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

```{r}
x <- as.data.frame(medium.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(medium.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(medium.mi3$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

```{r}
x <- as.data.frame(large.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(large.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(large.mi3$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

## Pred

```{r}
x <- as.data.frame(small.mi1$pred) %>% rename("pred" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

## Look at Coparenting Items 7 and 8

```{r}
here <- c("Coparent_item7_Mother_Year5", "Coparent_item8_Mother_Year5",
          "Coparent_item7_Father_Year5", "Coparent_item8_Father_Year5")
mice_df(large.mi1, vs = here)
```


# Emo Reg

# Load all mids objects

```{r}
# Small
load("0_data/processed_emoreg/small.mi1/small.mi1.Rdata")
small.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_emoreg/small.mi2/small.mi2.Rdata")
small.mi2 <- mi.res
rm(mi.res)

# Medium
load("0_data/processed_emoreg/medium.mi1/medium.mi1.Rdata")
medium.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_emoreg/medium.mi2/medium.mi2.Rdata")
medium.mi2 <- mi.res
rm(mi.res)

# Large
load("0_data/processed_emoreg/large.mi1/large.mi1.Rdata")
large.mi1 <- mi.res
rm(mi.res)

load("0_data/processed_emoreg/large.mi2/large.mi2.Rdata")
large.mi2 <- mi.res
rm(mi.res)
```

## Logged Events

### Small

```{r}
x <- small.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- small.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
```

### Medium

```{r}
x <- medium.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- medium.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
```

### Large

```{r}
x <- large.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))

x <- large.mi1$loggedEvents %>%
  filter(grepl('collinear', out))
length(table(x$dep))
```

## Meth

```{r}
x <- as.data.frame(small.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(small.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

```{r}
x <- as.data.frame(medium.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(medium.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

```{r}
x <- as.data.frame(large.mi1$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])

x <- as.data.frame(large.mi2$method) %>% rename("meth" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```

## Pred

```{r}
x <- as.data.frame(small.mi1$pred) %>% rename("pred" = 1)
x <- as.data.frame(x[!apply(x == "", 1, all),])
```
